Detect database dead-lock and raise exception appropriately.
Let API functions take a parameter that can be either a tag or an ID
for albums (and a hash or an ID for images). Renamed
ReservedAlbumTagError exception to ReservedAlbumError so that it can
apply to both cases.

Moved verifyAlbumTag from Shelf to a module-global function (and
renamed to verifyValidAlbumTag). Added parameter to choose whether to
check for ID-like tags.

Command-line client: Detect argument errors in command-line client.
Revised help text a bit. Prepared for the "add" command.