The Grand Kofoto Unicodification:
All database strings are now stored as UTF-8. All strings passed as
arguments to Shelf methods are now expected to be in Unicode, and all
returned strings are in Unicode.

Text printed to standard output and standard error are encoded
according to the user's locale settings.

Added XML declarations to generated XHTML files so that character
encoding can be specified. The encoding defaults to the codeset of the
locale, but can be overridden with the new --gencharenc option flag.

Implemented a wrapper for sqlite.Connection called
_UnicodeConnectionDecorator that overrides the cursor method of the
decorated Connection; an _UnicodeCursorDecorator instance is returned
instead. The _UnicodeCursorDecorator wraps a database cursor and
essentially does two things:

    1. Assures that all SQL string arguments are in Unicode (an
       exception is raised otherwise) and then encodes them into
       UTF-8. (The purpose of not touching non-Unicode strings even
       with a long stick is to avoid storing non-UTF-8 data in the
       database due to bugs, e.g. forgotten conversions in some code.)
    2. Decodes database strings from UTF-8 into Unicode.

A _UnicodeCursorDecorator instance is also an iterable, so the
_cursoriter function was removed.

The locals()-in-SQL-query idiom is no longer used in any query, since
the _UnicodeCursorDecorator chokes in non-Unicode string values in the
supplied dictionary (just like it's supposed to).