Implemented caching of albums, images, categories, attributes AND MORE!
New module: kofoto.sets, a subset of the sets module in Python 2.3 but
compatible with Python 2.2. It may be removed in the future when
Python 2.3 has been around for a while.

New module: kofoto.dag that implements the DAG datatype.

New module: kofoto.cachedobject, containing a kind of singleton-like
class but with ability of invalidating the instance.

Removed kofoto.sqlset since it isn't used anymore.