Rewrote the image cache interface. Image location and timestamp are now used when looking up images in the cache instead of the MD5 checksum. This makes it possible to cache unregistered images too. Made the cleanup method remove cached images for which the original images don't exist anymore. The sizes of the cached images are no longer considered when cleaning.
Moved image width and height attributes to the image table. Added
Image.getSize method (and a testcase).

The image cache interface change requires knowledge of the timestamp
(modification time) in some cases, so it's now stored per image in the
shelf. New methods: Image.contentChanged and Image.locationChanged.
They supersede the old Image.setHash and Image.setLocation methods
(which were removed). In addition to checksum, contentChanged also
records modification time and image size.

It is now possible to call Shelf.getImage to look up an image given a
location (instead of a checksum) if the new optional
identifyByLocation argument is true. The new exception
MultipleImagesAtOneLocationError is thrown if there are several
matching images. Wrote testcases for the methods.

Bumped shelf format version to 2 and wrote autoupgrade code.

Store images in generated albums in per-month directories instead of
the old two-level checksum structure. This makes manual browsing of
the image directory much nicer.

Removed the obsolete cachedir module.

Removed cycle in shelftests.TestShelfFixture since it made the test
suite fail on Windows. The cycle prevented garbage collection of the
SQLite connection so that the database file was kept open, which made
Windows unhappy in the tearDown method where the shelf file is
deleted.

Improved update-contents and update-locations commands in the
commandline client. They now take a number of paths, which are
traversed recursively.

Shelf.getAllAttributeNames was called before Shelf.begin in gkofoto.
Fixed.