Wrote a webkofoto wrapper program that imports webkofoto.py as a module in a kofotoweb package.
Added support for specifying the port to bind to.

Removed webkofoto.cfg; it is now generated on-the-fly by webkofoto.

Build and install webkofoto on POSIX systems, since it seems tricky to
do this automatically on Windows.

Renamed kofoto.css to webkofoto.css.

Set svn:ignore and svn:eol-style properties where appropriate.

Mention CherryPy in INSTALL.